Volume unities

+++ calculator/mass_test.go
@@ -4,14 +4,14 @@ import "testing"
 
 func TestConvertMass(t *testing.T) {
 
-	kilograms := float64(1)
+	kilograms := 1.0
 	grams := ConvertMass(kilograms).FromKilogram().ToGram()
 
 	if grams != 1000 {
 		t.Error("Unexpected conversion")
 	}
 
-	milligrams := float64(1000)
+	milligrams := 1000.0
 	kilograms = ConvertMass(milligrams).FromMilligram().ToKilogram()
 
 	if kilograms != 0.001 {
